Added support for multiplayer.
Added gamepad support to the multiplayer game.

Changed to joystick mouse emulation, the right stick moves the mouse and
left trigger is mouse button 2 and left trigger mouse button 1. This was
necessary because for multiplayer and possibly game mods, it will send
mouse controls to the VM for UI and possible for some additional game
controls.

Added a temporary debug console command to switch between mouse
emulation and direct joystick control, in case I run into issues and
need to compare.