UI hack to display and change to more than 8 resolutions.
This UI hack is necessary because the UI code for multiplayer is in the
VM. In theory this should work for Mods that allow resolution changing
but it is untested! In fact only JA SP and JA MP are tested.

This UI hack detects when the UI asks for resolution strings and returns
a string from the real resolution/vidmode list. It detects when the
requested string changes, as an increment or decrement, to update the
string.

r_mode is now a UI only variable (r_video_mode is its general
replacement) and it was necessary to hijack cvar to detect when the UI
for video settings is first opened (r_mode is read) or the resolution
mode is applied (r_mode is changed).