Stereoscopic Unbuffered
This is an attempt at using a single buffer to do stereoscopic 3d. Still
has some bugs, some are general to the engine, and some are likely
related to trying to use the viewport to render to the same buffer. I
did one fix to the glow effect.

This implementation is not easy and too invasive, so this is here for
future reference.