Fix for Single Player, Left/Right image sync issues.
This fixes that the left and right images become desynced (the right
image is slightly more in the future). Most noticable in debug, when
performance is lower.

Unfortunately this fix only works for single player since a lot of the
code in multiplayer inside of the VM. I may need to come up with a more
general solution, and if I do I may revert this and replace it with that
solution.