Adds Depth! Realy Stereoscopic 3D!
What was before was a "pop out effect" only, all objects pop out of the
screen. This change allows things to look like they are far away.

2 new settings r_stereoDisplayWidth, r_stereoEyeDistance. These are to
be set to the width of your monitor/TV and the distance between your
eyes. With these, things in the distance will make your eyes parallel
instead of converged.