Fix to prevent displaying a previous frame.
For a brief moment an old frame is shown. This occurs because nothing is
drawn to the screen but a SwapBuffer (WIN_Present) is still done.

It's a bug that is hard to notice, only known occurance (but it could
happen elsewhere) is at the end of the load screen.