Fix for opening fmv getting stretched after transition.
This was caused by a previous fix for dissolves. Basically
RE_InitDissolve can be called at various times. Sometimes after a
GL_Present and sometimes before. If it's called before, it will capture
the last screen, if it is called after it will capture both the left and
right images, so it needs to stretch out one image. In the case of this
bug, it was always stretching, which fixed loading screen transitions
but during an FMV to FMV transition of the 'New Game' intro movie, it is
called before GL_Present and was unnecessarily being stretched.