UI 3D Elements in sterescopic
This adds the stereoscopic 'effect' to 3D elements in the UI. It is a
little more complex because some 3D elements, like the talking heads
between missions, have an offset that made them pop way out of the
screen.

New cvar that is for the distance you sit from the 3d display/tv, so
perspective projection looks correct to your eyes.