Initial functional VR stuff
This is not OpenJK redone in VR, rather this is an experiment in merely
augmenting games in VR. These kinds of games make people sick/nausiated.
So this treats the game as running on an advanced stereoscopic 3d tv
with eye tracking.

A few things don't work. The beginning movie scrolling text doesn't
work. Transition dissolves don't work. There is a weird distortion
effect that is painful on the eyes, may not be working or may be
improved. I may disable dissolves and distortions.

I also wanted to optimize stencil (scissor) viewports but it seems the
first time it is drawn, either an opengl state is preventing it from
drawing or it is being cleared away at some point. It doesn't seem to be
a clear.