Distortion Effect
Looks like code to work with older hardware that doesn't support NPOT
textures (Non-Power Of Two) was causing extra scaling/distortion to
occur dependent on resolution. but we are doing VR and we likely have
new hardware that supports NPOT, so I got rid of it. I also scaled down
the distortion effect, and adjusted for aspect ratio differences from
640/480.