3D to 3D Dissolves
Added support for capturing and replaying previous render command lists
to an offscreen buffer with VR head tracking. This allows for dissolving
between one 3d scene to another, which is useful since movies are drawn
far away now.

This is also an experiment in double buffering render commands, which
may be part of a bigger change to get multithreaded rendering working.