Fix for a black flash in the opening FMV.
So with the opening FMV, there is a transition from one FMV to another.
The other FMV starts off with a black frame which is why the screen
flashes black. The dissolve transition tries to draw the previous FMVs
last frame, but the new FMV has replaced it. So instead, we try to avoid
allocating the same handle (which ends up as the same texture) avoid the
new FMV replacing the last frame of the old FMV.