Failed attempt at Multithreading
I've given up on multithreading but wanted to push it to the repo
anyway, in case I want to pick it up again, or wanted to borrow code I
created here for something else.

Basically I learned a number of things. First there is a lot of floating
globals used in rendering, there would need to be a lot of isolation of
render code to make sure those globals are not modified/accessed
simultaniously. In the end, it might be better to capture a mix of raw
opengl calls or new custom render commands, in addition to existing
render commands (so they can be updated to new HMD tracking info).

Performance of loading screens is still bad, even though it is
multithreaded. This might be solvable by simply reducing how many
textures are loaded between redraws, there could also be some other
techniques that allow for textures to be loaded on separate threads
under the same context, or to share textures between two contexts.

If I were to pick this up again, I would try and only use it for doing
the loading screens instead, but given the bad performance I would
probably not pick it up again.